[UPDATE] rename UserStore to userStore
