+++ src/screens/LoadingScreen.tsx
@@ -2,7 +2,7 @@ import { inject, observer } from 'mobx-react';
 import * as React from 'react';
 import { NavigationScreenProp, SafeAreaView } from 'react-navigation';
 import Loading from '../components/Loading';
-import { UserStore } from '../store/UserStore';
+import { UserStore } from '../store/userStore';
 
 interface LoadingScreenProps {
   navigation: NavigationScreenProp<{}>;
